% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ramp_d.R
\name{colorRamp_d}
\alias{colorRamp_d}
\title{colourRamp_d}
\usage{
colorRamp_d(
  colors,
  n,
  bias = 1,
  space = c("rgb", "Lab"),
  interpolate = c("linear", "spline"),
  alpha = FALSE
)
}
\arguments{
\item{colors}{internal}

\item{n}{internal}

\item{bias}{internal}

\item{space}{internal}

\item{interpolate}{internal}

\item{alpha}{internal}
}
\description{
Variant of \code{colorRamp} for sequential use of colours for discrete scales. \href{https://stackoverflow.com/questions/61674217/custom-discrete-color-scale-in-ggplot-does-not-respect-order}{Thank you linog}.
}
