% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorRampPalette_d.R
\name{colorRampPalette_d}
\alias{colorRampPalette_d}
\title{colorRamPalette_d}
\usage{
colorRampPalette_d(colors, ...)
}
\arguments{
\item{colors}{internal}

\item{...}{internal}
}
\description{
Variant of \code{colorRampPalette} for sequential use of colours for discrete scales. \href{https://stackoverflow.com/questions/61674217/custom-discrete-color-scale-in-ggplot-does-not-respect-order}{Thank you linog}.
}
