% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_cholesterol}
\alias{data_cholesterol}
\title{data_cholesterol}
\format{
tibble:	30 obs. of  3 variables:

\describe{
\item{Hospital}{Factor with 5 levels (a-e), representing different hospitals where subjects were recruited.}
\item{Subject}{A factor with 25 levels denoting individuals on whom measurements were made twice.}
\item{Treatment}{A factor with 2 levels indicating when measurements were made, i.e. before and after drug.}
\item{Cholesterol}{Numerical dependent variable indicating measured doubling time in min.}
}
}
\usage{
data_cholesterol
}
\description{
An example dataset on measurements of blood cholesterol levels measured in 5 subjects measured before and after receiving a Drug. Five patients each were recruited at 5 hospitals (a-e), so that there are 25 different subjects (1-25) measured twice.  Data are from \href{https://microimmunostats.netlify.app}{Micro/Immuno Stats}
}
\keyword{datasets}
