% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_doubling_time}
\alias{data_doubling_time}
\title{Doubling time of E.coli measured by 10 students three independent times.}
\format{
tibble:	30 obs. of  3 variables:

\describe{
\item{Student}{Factor with 10 levels, representing different students.}
\item{Experiment}{A factor with 3 levels representing independent experiments.}
\item{Doubling_time}{Numerical dependent variable indicating measured doubling time in min.}
}
}
\usage{
data_doubling_time
}
\description{
An example dataset showing measurements of \emph{E. coli} doubling times (in min) measured by 10 different students in 3 independent experiments each. Note that Experiments are just called Exp1-Exp3 even though Exp1 of any of the students are not connected in anyway - this will confuse R! Data are from \href{https://microimmunostats.netlify.app}{Micro/Immuno Stats}
}
\keyword{datasets}
