% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_t_pdiff}
\alias{data_t_pdiff}
\title{Matched data from two groups where difference between them is consistent.}
\format{
data.frame:	20 obs. of  3 variables:

\describe{
\item{Subject}{Factor with 10 levels, denoted by capital letters, representing individuals or subjects.}
\item{Condition}{A fixed factor with 2 levels: "Untreated" & "Treated".}
\item{Mass}{Numerical dependent variable indicating body mass of mice}
}
}
\usage{
data_t_pdiff
}
\description{
An example dataset for paired difference Student's \emph{t} test. These are bodyweight (Mass) in grams of same mice left untreated or treated, which are two groups to be compared. The data are in a longtable format, and the two groups are levels within the factor "Condition". The Subject column lists ID of matched mice that were measured without and with treatment. These data are from \href{https://www.science.org/doi/10.1126/scisignal.aat6903}{Sanchez-Garrido \emph{et al}, Sci Signal, 2018, DOI: 10.1126/scisignal.aat6903}.
}
\keyword{datasets}
