% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_t_pratio}
\alias{data_t_pratio}
\title{Matched data from two groups where ratio between them is consistent.}
\format{
data.frame:	66 obs. of  3 variables:

\describe{
\item{Genotype}{Factor with 2 levels, representing genotypes to be compared ("WT" & "KO").}
\item{Experiment}{A random factor with 33 levels representing independent experiments, denoted as "Exp_1", "Exp_2"...}
\item{Cytokine}{Numerical dependent variable indicating cytokine measured by ELISA.}
}
}
\usage{
data_t_pratio
}
\description{
An example dataset for paired ratio Student's \emph{t} test. These are Cytokine measurements by ELISA (in ng/ml) from 33 independent in vitro experiments performed on two Genotypes that we want to compare. The data are in a longtable format, and the two groups are levels within the factor "Genotype". The Experiment column lists ID of matched experiments.
}
\keyword{datasets}
