% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graf_colours.R
\name{get_graf_colours}
\alias{get_graf_colours}
\title{Get graf internal}
\usage{
get_graf_colours(...)
}
\arguments{
\item{...}{internal}
}
\description{
Function to make grafify colour scheme. \href{https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}{Thank you Dr Simon}.
}
