% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_summary.R
\name{get_summary}
\alias{get_summary}
\title{Get numeric summary grouped by factors}
\usage{
get_summary(Ycol, ByXcol, data)
}
\arguments{
\item{Ycol}{name of one column or a vector of column names containing the numerical variable to be summarised.}

\item{ByXcol}{name of one column or a vector of column names containing the grouping factors}

\item{data}{name of the data table.}
}
\value{
this function takes in a data.frame or tibble and returns a data.frame or tibble.
}
\description{
This is a wrapper around \code{\link{aggregate}} function in base R to obtain mean, median, standard deviation and count for quantiative variable(s) grouped by one or more factors.
}
\examples{
get_summary(Ycol = "cty",
ByXcol = c("fl", "drv"),
data = mpg)

}
