% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf_colour_palettes_fun.R
\name{graf_col_palette}
\alias{graf_col_palette}
\title{Call palettes for scale & fill}
\usage{
graf_col_palette(palette = "okabe_ito", reverse = FALSE, ...)
}
\arguments{
\item{palette}{internal}

\item{reverse}{internal}

\item{...}{additional parameters}
}
\value{
This generates required number of sequential colours from the chosen grafify palette when called by scale functions of  ggplot2.
}
\description{
Call palettes for scale & fill
}
