% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2way_data.R
\name{make_2way_data}
\alias{make_2way_data}
\title{Make one-way or two-way independent group or randomised block design data.}
\usage{
make_2way_data(Group_1_means, Group_2_means, Num_obs, Residual_SD)
}
\arguments{
\item{Num_obs}{a single numeric value indicating the number of independent measurements, i.e. levels within the random factor Experiment}

\item{Residual_SD}{a single numeric value indicating residual SD in the model}

\item{Group1_means}{a vector with means of each level of the first fixed factor (FixFac_X1) measured within Group 1}

\item{Group2_means}{only for \code{make_2way_data} and \code{make_2way_rb_data}: a vector with mean(s) of each level of FactorX2 measured within Group 2}
}
\value{
This function produces a \code{data.frame} object
}
\description{
The \code{\link{make_1way_data}}, \code{\link{make_1way_rb_data}}, \code{\link{make_2way_data}} and \code{\link{make_2way_rb_data}} functions generate independent or randomised block (rb) design data of one-way or two-way designs.
}
\details{
Random variates from the normal distribution based on user provided mean and SD provided are generated. For independent designs, the \code{Residual_SD} argument is used to set expected residual SD from the linear model. Exp_SD is used to set experiment-to-experiment SD, that will be assigned to the random factor for rb designs.

Num_obs sets the number of independent measurements per group.

For one-way designs, the user provides Group_means as a vector. Number of levels are recognised based on number of means.
For two-way designs, two vectors are to be provided by the user containing means of levels of a second factor. Number of means in both vectors should be the same. These functions can only handle balanced designs, i.e. same number of observations in all groups.

The output is a data frame with one or two columns denoting the fixed factor with levels that match the number of means entered. For rb data, the column for RandFac denotes levels of the blocking factor.
The quantitative response variables are in the numeric Values column.
}
\examples{
#Basic usage with two levels within FactorX2, 20 observations in each group, with residual SD 15

two_independent_tab <- make_2way_data(c(100, 20), c(200, 300), 20, 15)

str(two_independent_tab)
head(two_independent_tab)

#Four levels with 5 observations and residual SD 5
two_independent_tab <- make_2way_data(c(100, 20, 1500, 20), c(150, 5, 1450, 25), 5, 5)
}
