% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_befafter_colors.R
\name{plot_befafter_colors}
\alias{plot_befafter_colors}
\title{Plot a before-after plot with lines joining colour-matched symbols.}
\usage{
plot_befafter_colors(
  data,
  xcol,
  ycol,
  groups,
  symsize = 3,
  symthick = 1,
  fontsize = 20,
  alpha = 1
)
}
\arguments{
\item{data}{a data table object, e.g. data.frame or tibble.}

\item{xcol}{name of the column containing the categorical variable to be plotted on the X axis.}

\item{ycol}{name of the column containing the quantitative Y values.}

\item{groups}{name of the column with the grouping variable to pass on to \code{geom_line}.}

\item{symsize}{size of symbols, default set to 3}

\item{symthick}{thickness of symbol border, default set to 1}

\item{fontsize}{parameter of \code{base_size} of fonts in \code{theme_classic}, default set to size 20.}

\item{alpha}{fractional transparency of symbols, default set to 1 (i.e. zero transparency, fully opaque)}
}
\value{
This function returns a \code{ggplot2} object.
}
\description{
The \code{\link{plot_befafter_colours}} and \code{\link{plot_befafter_shapes}} are for graphing matched data joined by lines. They take X and Y variables along with a grouping factor (e.g. matched subjects or experiments etc.) and plot symbols matched by colour or shape.
}
\details{
Note that \code{ggplot} only allows 25 types of shapes, and \code{\link{plot_befafter_shapes}} function works when there are fewer than 25 matched observations.
Use \code{scale_colour_brewer} or \code{scale_colour_viridis} or related to get a spectrum of matched colours by the grouping factor.

Consider using \code{facet_wrap} or \code{facet_grid} to reduce crowding on a single plot and/or plot additional dimensions.
}
\examples{
#Basic usage with Treatment as the X variable with Subject as the grouping variable that indicates which points to join by lines
plot_befafter_colours(Chol, Treatment, Cholesterol, Subject)

#Additional layers are possible
plot_befafter_colours(Chol, Treatment, Cholesterol, Subject)+
   labs(title = "Plot with scatter plot, bars (mean) & SD")+
   scale_color_viridis_d()+scale_fill_viridis_d()+
   facet_wrap("Hospital")
}
