% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotbox.R
\name{plot_dotbox}
\alias{plot_dotbox}
\title{Plot a dotplot on a boxplot with two variables.}
\usage{
plot_dotbox(
  data,
  xcol,
  ycol,
  dotsize = 1.5,
  dotthick = 1,
  fontsize = 20,
  alpha = 1,
  ColPal = "all_grafify",
  ColRev = F,
  TextXAngle = 0
)
}
\arguments{
\item{data}{a data table object, e.g. data.frame or tibble.}

\item{xcol}{name of the column to plot on X axis. This should be a categorical variable.}

\item{ycol}{name of the column to plot on quantitative Y axis. This should be a quantitative variable.}

\item{dotsize}{size of dots relative to binwidth used by \code{geom_dotplot}. Default set to 1.5, increase/decrease as needed.}

\item{dotthick}{thickness of dot border (\code{stroke} parameter of \code{geom_dotplot}), default set to 1}

\item{fontsize}{parameter of \code{base_size} of fonts in \code{theme_classic}, default set to size 20.}

\item{alpha}{fractional opacity of boxplot, default set to 1 (i.e. maximum opacity & zero transparency)}

\item{ColPal}{grafify colour palette to apply, default "all_grafify"; alternatives: "okabe_ito", "bright", "pale", "vibrant", "contrast", "muted" "dark", "light".}

\item{ColRev}{whether to reverse order of colour choice, default F (FALSE); can be set to T (TRUE)}

\item{TextXAngle}{orientation of text on X-axis; default 0 degrees. Change to 45 or 90 to remove overlapping text}
}
\value{
This function returns a \code{ggplot2} object on which additional geometries etc. can be added.
}
\description{
This function takes a data table, X and Y variables, and plots a graph with a dotplot and boxplot using \code{ggplot}.
}
\details{
The function uses \code{\link[ggplot2]{geom_boxplot}} and \code{\link[ggplot2]{geom_dotplot}} geometries.
Note that \code{\link{geom_boxplot}} option for outliers is set to \code{outlier.alpha = 0}.
The X variable is mapped to the \code{fill} aesthetic in both boxplot and dotplot, and its colour can be changed using \code{scale_fill_brewer} or any \code{scale_fill...} option. The size of dots can be adjusted using the parameter, which is \code{dotsize = 1} by default.

This function is related to \code{\link{plot_scatterbar_sd}}, \code{\link{plot_dotbar_sd}} and \code{\link{plot_dotviolin}}.
}
\examples{

#Basic usage requires a data table and X & Y variables#'
plot_dotbox(data_cholesterol, Treatment, Cholesterol)

#Transformations of Y variable are possible as follows
#' plot_dotbox(data_cholesterol, Treatment, log(Cholesterol))

#Additional ggplot layering is possible
plot_dotbox(data_cholesterol, Treatment, Cholesterol, dotsize = 2)+
   labs(title = "Plot with scatter dots & boxplot")+
   scale_colour_grafify()+
   facet_wrap("Hospital")
}
