% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gam_models.R
\name{plot_gam_qq}
\alias{plot_gam_qq}
\title{Plot model diagnostics for generalised additive models}
\usage{
plot_gam_qq(Model, ...)
}
\arguments{
\item{Model}{a model of class \code{gam} fitted with \code{ga_model} or the \code{mgcv} package.}

\item{...}{additional arguments to pass to \code{\link[gratia]{appraise}}.}
}
\value{
This function returns an object of classes "ggplot", "gg" and "patchwork".
}
\description{
This is a clone of the \code{\link[gratia]{appraise}} function in the \code{gratia} package and will plot 4 diagnostic plots when given a generalised additive model fitted with \code{\link{ga_model}}. It creates graphs that use \code{grafify} colours and \code{theme_classic()}.
}
