% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gam_models.R
\name{plot_gam_smooth}
\alias{plot_gam_smooth}
\title{Plot fitted smooths of generalised additive models}
\usage{
plot_gam_smooth(Model, ...)
}
\arguments{
\item{Model}{a model of class \code{gam} fitted with \code{ga_model} or the \code{mgcv} package.}

\item{...}{additional arguments to pass to \code{\link[gratia]{draw}}.}
}
\value{
This function returns an object of classes "ggplot", "gg" and "patchwork
}
\description{
This is a clone of the \code{\link[gratia]{draw}} function in the \code{gratia} package. It will plot smooths for all levels of all fixed factors in a generalised additive model fitted with \code{\link{ga_model}}. It creates graphs that use \code{grafify} colours and \code{theme_classic()}.
}
