% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatterbox.R
\name{plot_scatterbox}
\alias{plot_scatterbox}
\title{Plot a scatter plot on a boxplot with two variables.}
\usage{
plot_scatterbox(
  data,
  xcol,
  ycol,
  symsize = 2.5,
  symthick = 1,
  jitter = 0,
  fontsize = 20,
  b_alpha = 1,
  s_alpha = 1,
  ColPal = "all_grafify",
  ColRev = F,
  TextXAngle = 0
)
}
\arguments{
\item{data}{a data table object, e.g. data.frame or tibble.}

\item{xcol}{name of the column to plot on X axis. This should be a categorical variable.}

\item{ycol}{name of the column to plot on quantitative Y axis. This should be a quantitative variable.}

\item{symsize}{size of symbols used by \code{geom_point}. Default set to 2.5, increase/decrease as needed.}

\item{symthick}{thickness of symbol border (\code{stroke} parameter of \code{geom_point}), default set to 1}

\item{jitter}{extent of jitter (scatter) of symbols, default is 0 (i.e. aligned symbols). To reduce symbol overlap, try 0.1-0.3 or higher.}

\item{fontsize}{parameter of \code{base_size} of fonts in \code{theme_classic}, default set to size 20.}

\item{b_alpha}{fractional opacity of boxplot, default set to 1 (i.e. maximum opacity & zero transparency)}

\item{s_alpha}{fractional opacity of symbols, default set to 1 (i.e. maximum opacity & zero transparency)}

\item{ColPal}{grafify colour palette to apply, default "all_grafify"; alternatives: "okabe_ito", "bright", "pale", "vibrant", "contrast", "muted" "dark", "light".}

\item{ColRev}{whether to reverse order of colour choice, default F (FALSE); can be set to T (TRUE)}

\item{TextXAngle}{orientation of text on X-axis; default 0 degrees. Change to 45 or 90 to remove overlapping text}
}
\value{
This function returns a \code{ggplot2} object on which additional geometries etc. can be added.
}
\description{
This function takes a data table, X and Y variables, and plots a graph with a scatter plot and boxplot using \code{ggplot}.
}
\details{
The function uses \code{\link[ggplot2]{geom_boxplot}} and \code{\link[ggplot2]{geom_point}} geometries.
Note that \code{\link{geom_boxplot}} option for outliers is set to \code{outlier.alpha = 0}.
The X variable is mapped to the \code{fill} aesthetic in both boxplot and symbols, and its colour can be changed using \code{ColPal} option. The size of symbols can be adjusted using \code{symsize} set to 1 by default.
Transparency of boxplot and symbols can be set independently with \code{b_alpha} and \code{s_alpha}, respectively.

Three types of plots are available for scatter/jitter symbols and either bars+SD, boxplot or violin plots: \code{\link{plot_scatterbar_sd}}, \code{\link{plot_scatterbox}} and \code{\link{plot_scatterviolin}}.
These are related to the three "dot" versions that use a different geometry for symbols: \code{\link{plot_scatterbox}}, \code{\link{plot_dotbar_sd}} and \code{\link{plot_dotviolin}}.
}
\examples{

#Basic usage requires a data table and X & Y variables#'
plot_scatterbox(data_cholesterol, Treatment, Cholesterol)

#with jitter
plot_scatterbox(data_cholesterol, Treatment, Cholesterol, jitter = 0.1)

}
