% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xy_CatGroup.R
\name{plot_xy_CatGroup}
\alias{plot_xy_CatGroup}
\title{Plot points on a quantitative X - Y plot & a categorical grouping variable.}
\usage{
plot_xy_CatGroup(
  data,
  xcol,
  ycol,
  CatGroup,
  symsize = 2,
  symthick = 1,
  fontsize = 20,
  s_alpha = 1,
  ColPal = "all_grafify",
  ColRev = F,
  TextXAngle = 0
)
}
\arguments{
\item{data}{a data table object, e.g. data.frame or tibble.}

\item{xcol}{name of the column with quantitative X variable}

\item{ycol}{name of the column with quantitative Y variable}

\item{CatGroup}{a categorical variable as grouping factor for colour of data points, should be a categorical variable for default colours to work.
Will be converted to \code{factor} if your column is numeric.}

\item{symsize}{size of point symbols, default set to 2}

\item{symthick}{thickness of symbol border, default set to 1}

\item{fontsize}{parameter of \code{base_size} of fonts in \code{theme_classic}, default set to size 20.}

\item{s_alpha}{fractional opacity of symbols, default set to 1 (i.e. maximum opacity & zero transparency)}

\item{ColPal}{grafify colour palette to apply, default "all_grafify"; alternatives: "okabe_ito", "bright", "pale", "vibrant", "contrast", "muted" "dark", "light".}

\item{ColRev}{whether to reverse order of colour choice, default F (FALSE); can be set to T (TRUE)}

\item{TextXAngle}{orientation of text on X-axis; default 0 degrees. Change to 45 or 90 to remove overlapping text}
}
\value{
This function returns a \code{ggplot2} object.
}
\description{
This function takes a data table, quantitative X and Y variables, and a and plots a graph with using \code{\link[ggplot2]{ggplot}}.
}
\details{
The function uses \code{\link[ggplot2]{geom_point}} with \code{size = 3}, which can be changed using the \code{symsize} argument.
The categorical \code{CatGroup} variable is mapped to the \code{fill} aesthetic of symbols, which receives the \code{scale_fill_grafify} default palette (\code{all_grafify}).
This plot is related to \code{\link{plot_xy_NumGroup}} which requires a numeric grouping factor.
When summary statistics (mean/median) are required, use \code{\link{plot_3d_scatterbar}}, \code{\link{plot_3d_scatterbox}} or \code{\link{plot_4d_scatterbox}}.
}
\examples{
#Basic usage with mtcars data
#The grouping factor gear is converted to factor automatically
plot_xy_CatGroup(mtcars, mpg, disp, gear)
}
