% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_graf.R
\name{scale_color_grafify}
\alias{scale_color_grafify}
\title{Scale_colour colour scheme}
\usage{
scale_color_grafify(palette = "all_grafify", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of the colour scheme. Default set to \code{palette = "all_grafify"}. Provide names as above in quotes.}

\item{reverse}{Whether the colour order should be reversed.}

\item{...}{Additional parameters for \code{scale_fill} or \code{scale_colour}}
}
\value{
ggplot scale_fill function
}
\description{
\code{grafify} internally includes colour-blind compatible schemes for fill and colour/color aesthetics. Note that these \strong{only} work for categorical variables. Use the brewer or viridis packages for numeric gradient scales.
}
\details{
The default for \code{scale_fill_grafify()}, \code{scale_colour_grafify()} or \code{scale_color_grafify()} is a list of 46 colours as part of \code{palette = "all_grafify"}.

Obviously, it is not recommended to use so many colours, but implementing this was easiest to prevent errors when using a lot of categorical variables.

There are six other schemes with 6-10 colours each, which are recommended. These can be called by naming the colour scheme using \code{palette = } argument.
Additional options include "okabe_ito", "bright", "pale", "muted", "dark" and "light". These are taken from \href{https://personal.sron.nl/~pault/#sec:qualitative}{Paul Taul},  \href{https://mikemol.github.io/technique/colorblind/2018/02/11/color-safe-palette.html}{Mike Mol} and \href{http://jfly.iam.u-tokyo.ac.jp/color/#pallet}{Okabe Ito}.
}
\examples{
#if p is a ggplot object produced by any of the `plot_` functions or `ggplot2`
p + scale_colour_grafify() #for point and scatter graphs
p + scale_colour_grafify() + scale_fill_grafify() #for dotbar, dotbox etc

#specify colour palette & reverse order
p + scale_colour_grafify(palette = "okabe_ito", reverse = T)

}
