% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_grafify_c.R
\name{scale_color_grafify_c}
\alias{scale_color_grafify_c}
\title{Scale_color continuous scheme}
\usage{
scale_color_grafify_c(reverse = FALSE, ...)
}
\arguments{
\item{reverse}{Whether the color order should be reversed.}

\item{...}{Additional parameters for \code{scale_fill} or \code{scale_color}}
}
\value{
ggplot scale_fill function
}
\description{
\code{grafify} internally includes color-blind compatible schemes for fill and color/color aesthetics.
Note that this scheme is \strong{only} for continuous variables and has one palette (\code{yellow_conti}) \href{http://personal.sron.nl/~pault/#sec:sequential}{modified from} the YlOrBr scheme from RColorBrewer.
}
\examples{
#basic usage on mtcars data with x and y quantitative axes
ggplot(mtcars, aes(x = mpg, y = disp))+
geom_point(aes(color= disp))+
scale_color_grafify_c()

}
