% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_mod_summary.R
\name{simple_mod_summary}
\alias{simple_mod_summary}
\title{Model summary from fit of a linear model to data.}
\usage{
simple_mod_summary(data, Y_value, Fixed_Factor, ...)
}
\arguments{
\item{data}{a data table object, e.g. data.frame or tibble}

\item{Y_value}{name of column containing quantitative (dependent) variable, provided within "quotes".}

\item{Fixed_Factor}{name(s) of categorical fixed factors (independent variables) provided as a vector if more than one or within "quotes".}

\item{...}{any additional arguments to pass on to \code{lm} or \code{anova}.}
}
\value{
This function returns the output of from \code{summary(lm())}.
}
\description{
This function uses \code{\link{lm}} to fit a linear model to data and provides the full \code{\link{summary}}. It requires a data table, one quantitative dependent variable and one or more independent variables. The model output can be used to extract coefficients and other information. If your experiment design has random factors, use the related function \code{\link{mixed_mod_summary}}.
}
\details{
This function is related to \code{link{simple_anova}} and \code{\link{simple_model}}.
}
\examples{
#Basic usage where the table Chol is passed with names of one variable within quotes

simple_lmod_summary(Chol, "Cholesterol", "Treatment")

#two way ANOVA with Treatment & Hospital as fixed factors
simple_lmod_summary(Chol, "Cholesterol", c("Treatment", "Hospital"))

}
