% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf_colour_palettes_default_fun.R
\name{graf_col_palette_default}
\alias{graf_col_palette_default}
\title{Call palettes for scale & fill with default colorRampPalette}
\usage{
graf_col_palette_default(palette = "all_grafify", reverse = FALSE, ...)
}
\arguments{
\item{palette}{internal}

\item{reverse}{internal}

\item{...}{additional parameters}
}
\value{
This generates required number of distant colours from the chosen grafify palette when called by scale functions of  ggplot2.
}
\description{
Call palettes for scale & fill with default colorRampPalette
}
