% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf_colour_list.R
\docType{data}
\name{graf_colours}
\alias{graf_colours}
\title{List of hexcodes of colours in grafify palettes}
\format{
An object of class \code{character} of length 64.
}
\usage{
graf_colours
}
\description{
To visualise these colours use \code{plot_grafify_palette}.
}
\keyword{datasets}
