% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf_palette_list.R
\docType{data}
\name{graf_palettes}
\alias{graf_palettes}
\title{List of palettes available in grafify package}
\format{
An object of class \code{list} of length 10.
}
\usage{
graf_palettes
}
\description{
To visualise these colours use \code{plot_grafify_palette}.
}
\keyword{datasets}
