% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot density distribution of data.}
\usage{
plot_density(
  data,
  ycol,
  Group,
  linethick = 1,
  fontsize = 20,
  alpha = 0.2,
  ColPal = "all_grafify",
  ColRev = F,
  TextXAngle = 0
)
}
\arguments{
\item{data}{a data table e.g. data.frame or tibble.}

\item{ycol}{name of the column containing the quantitative variable whose density distribution is to be plotted}

\item{Group}{name of the column containing a categorical grouping variable}

\item{linethick}{thickness of symbol border, default set to 1}

\item{fontsize}{parameter of \code{base_size} of fonts in \code{theme_classic}, default set to size 20.}

\item{alpha}{fractional opacity of symbols, default set to 0.2 (i.e. 20\% opacity)}

\item{ColPal}{grafify colour palette to apply, default "all_grafify"; alternatives: "okabe_ito", "bright", "pale", "vibrant", "contrast", "muted" "dark", "light".}

\item{ColRev}{whether to reverse order of colour choice, default F (FALSE); can be set to T (TRUE)}

\item{TextXAngle}{orientation of text on X-axis; default 0 degrees. Change to 45 or 90 to remove overlapping text}
}
\value{
This function returns a \code{ggplot2} object.
}
\description{
This function takes a data table, X and a Grouping variable (if available) and plots a density graph using \code{\link[ggplot2]{ggplot}}.
}
\details{
The function uses \code{\link{geom_density}}).
Note that the function requires the quantitative Y variable first, and groups them based on an X variable.
The Group variable is mapped to the \code{fill} and \code{colour} aesthetics in \code{geom_density}.
ColPal & ColRev options are applied to both \code{fill} and \code{colour} scales.
}
\examples{
#Basic usage
plot_density(data_t_pratio, Cytokine, Genotype)+scale_x_log10()
}
