% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotbar_sd.R
\name{plot_dotbar_sd}
\alias{plot_dotbar_sd}
\title{Plot a dotplot on a bar graph with SD error bars with two variables.}
\usage{
plot_dotbar_sd(data, xcol, ycol, dotsize = 1)
}
\arguments{
\item{data}{a data table object, e.g. data.frame or tibble.}

\item{xcol}{name of the column to plot on X axis. This should be a categorical variable.}

\item{ycol}{name of the column to plot on quantitative Y axis. This should be a quantitative variable.}

\item{dotsize}{size of dots relative to binwidth used by \code{\link[ggplot2]{geom_dotplot}}. Default set to 1, increase/decrease as needed.}
}
\value{
This function returns a \code{ggplot2} object on which additional geometries etc. can be added.
}
\description{
This function takes a data table, X and Y variables, and plots a graph with a dotplot and bars using \code{\link[ggplot2]{ggplot}}.
}
\details{
The function uses \code{\link[ggplot2]{stat_summary}} with \code{geom = "bar"}, and \code{\link[ggplot2]{geom_dotplot}} geometries.
Standard deviation (SD) is plotted through \code{\link[ggplot2]{stat_summary}} calculated using \code{\link[ggplot2]{mean_sdl}} from the \code{ggplot2} package, and 1x SD is plotted (\code{fun.arg = list(mult = 1)}.
The X variable is mapped to the \code{fill} aesthetic in both bar and dotplot, and its colour can be changed using \code{scale_fill_brewer} or any \code{scale_fill...} option. The size of dots can be adjusted using the parameter, which is  \code{dotsize = 1} by default.

This function is related to \code{\link{plot_scatterbar_sd}}, \code{\link{plot_dotbar_sd}} and \code{\link{plot_dotviolin}}.
}
\examples{
#Basic usage requires a data table and X & Y variables#'
plot_dotbar_sd(Chol, Treatment, Cholesterol)

#Transformations of Y variable are possible as follows
#' plot_dotbar_sd(Chol, Treatment, log(Cholesterol))

#Additional ggplot layering is possible
plot_dotbar_sd(Chol, Treatment, Cholesterol, dotsize = 2)+
   labs(title = "Plot with scatter dots & boxplot")+
   scale_color_viridis_d()+scale_fill_viridis_d()+
   facet_wrap("Hospital")
}
