% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotviolin.R
\name{plot_dotviolin}
\alias{plot_dotviolin}
\title{Plot a dotplot on a violin plot with two variables.}
\usage{
plot_dotviolin(
  data,
  xcol,
  ycol,
  dotsize = 1,
  trim = T,
  scale = "area",
  fontsize = 20
)
}
\arguments{
\item{data}{a data table object, e.g. data.frame or tibble}

\item{xcol}{name of the column to plot on X axis. This should be a categorical variable.}

\item{ycol}{name of the column to plot on quantitative Y axis. This should be a quantitative variable.}

\item{dotsize}{size of dots relative to \code{binwidth} used by \code{geom_dotplot}. Default set to 1, increase/decrease as needed.}

\item{trim}{set whether tips of violin plot should be trimmed at high/low data. Default \code{trim = T}, can be changed to F.}

\item{fontsize}{parameter of \code{base_size} of fonts in \code{theme_classic}, default set to size 20.}
}
\value{
This function returns a \code{ggplot2} object on which additional geometries etc. can be added.
}
\description{
This function takes a data table, X and Y variables, and plots a graph with a dotplot and violinplot using \code{ggplot}.
}
\details{
The function uses \code{\link[ggplot2]{geom_violin}} and \code{\link[ggplot2]{geom_dotplot}} geometries.
Note that the \code{\link{geom_violin}} options are set as follows: \code{scale = "area", draw_quantiles = c(0.25, .5, .75)}. The \code{trim = T} set by default can be changed when calling the function.
The X variable is mapped to the \code{fill} aesthetic in both violinplot and dotplot, and its colour can be changed using \code{scale_fill_brewer} or any \code{scale_fill...} option. The size of dots can be adjusted using the parameter, which is \code{dotsize = 1} by default.

This function is related to \code{\link{plot_scatterbar_sd}}, \code{\link{plot_dotbar_sd}} and \code{\link{plot_dotviolin}}.
}
\examples{

#Basic usage requires a data table and X & Y variables#'
plot_scatterviolin(Tab_doublings, Student, Doubling_time)

# or with trim = F
plot_scatterviolin(Tab_doublings, Student, Doubling_time, trim = F)

#Additional ggplot layering is possible
plot_scatterviolin(Chol, Treatment, Cholesterol)+
   scale_fill_viridis_d()+
   scale_colour_viridis_d()
}
