% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grafify_palette.R
\name{plot_grafify_palette}
\alias{plot_grafify_palette}
\title{See grafify colour palettes}
\usage{
plot_grafify_palette(
  palette = c("okabe_ito", "all_grafify", "bright", "contrast", "dark", "fishy", "kelly",
    "light", "muted", "pale", "r4", "safe", "vibrant", "OrBl_div", "PrGn_div",
    "blue_conti", "grey_conti", "yellow_conti"),
  fontsize = 14,
  ...
)
}
\arguments{
\item{palette}{name of grafify palettes: "okabe_ito", "vibrant, "bright", "pale", "muted", "dark", "light", "contrast" or "all_grafify".}

\item{fontsize}{font size.}

\item{...}{any additional parameters to pass to \code{plot_bar_sd}}
}
\value{
This function returns a \code{ggplot2} object of class "gg" and "ggplot".
}
\description{
This simple function allows quick visualisation of colours in grafify palettes and their hex codes. It uses \code{plot_bar_sd} and some arguments are similar and can be adjusted.
}
\examples{
plot_grafify_palette("pale")
plot_grafify_palette("contrast")

}
