% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point_sd.R
\name{plot_point_sd}
\alias{plot_point_sd}
\title{Plot a point as mean with SD error bars using two variables.}
\usage{
plot_point_sd(data, xcol, ycol, size = 2, width = 0.2)
}
\arguments{
\item{data}{a data table object, e.g. data.frame or tibble.}

\item{xcol}{name of the column with X variable}

\item{ycol}{name of the column with quantitative Y variable}
}
\value{
This function returns a \code{ggplot2} object.
}
\description{
This function takes a data table, X and Y variables, and plots a graph with a summary point showing the mean of data using \code{\link[ggplot2]{ggplot}}.
}
\details{
The function uses \code{\link[ggplot2]{stat_summary}} with \code{geom = "point"} with \code{size = 3}.
Standard deviation (SD) is plotted through \code{\link[ggplot2]{stat_summary}} calculated using \code{\link[ggplot2]{mean_sdl}} from the \code{ggplot2} package, and 1x SD is plotted (\code{fun.arg = list(mult = 1)}.
The X variable is mapped to the \code{colour} aesthetic in the point geometry. Colour can be changed using \code{scale_colour_brewer} (or related) options.

You are instead encouraged to show all data using the following functions: \code{\link{plot_scatterbar_sd}}, \code{\link{plot_dotbox}}, \code{\link{plot_dotbar_sd}} and \code{\link{plot_dotviolin}}.
}
\examples{
#Basic usage
plot_point_sd(Tab_doublings, Student, Doubling_time)

#Additional layers can be added
plot_point_sd(Tab_doublings, Student, Doubling_time)+
   labs(title = "Plot with dot (mean) & SD")+
   scale_color_viridis_d()

}
