% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qqline.R
\name{plot_qqline}
\alias{plot_qqline}
\title{Plot quantile-quantile (QQ) graphs from data.}
\usage{
plot_qqline(data, ycol, xcol, symsize = 3, fontsize = 20)
}
\arguments{
\item{data}{a data table e.g. data.frame or tibble.}

\item{ycol}{name of the column containing the quantitative variable whose distribution is to be plotted}

\item{xcol}{name of the column containing a categorical variable}

\item{symsize}{size of symbols, default set to 3}

\item{fontsize}{parameter of \code{base_size} of fonts in \code{theme_classic}, default set to size 20.}
}
\value{
This function returns a \code{ggplot2} object.
}
\description{
This function takes a data table, X and Y variables, and plots a QQ graph using \code{\link[ggplot2]{ggplot}}.
}
\details{
The function uses \code{stat_qq} and \code{stat_qq_line} geometries (get help with \code{?stat_qq_line}).
Note that the function requires the quantitative Y variable first, and groups them based on an X variable.
The X variable is mapped to the \code{colour} aesthetic in both \code{stat_qq} and \code{stat_qq_line}, and its colour can be changed using \code{scale_colour_brewer} or any \code{scale_colour...} option.
}
\examples{
#Basic usage
plot_qqline(Chol, Cholesterol, Treatment)

Additional layers are possible
plot_qqline(Chol, Cholesterol, Treatment)+
   facet_wrap("Treatment")
}
