% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qqline.R
\name{plot_qqline}
\alias{plot_qqline}
\title{Plot quantile-quantile (QQ) graphs from data.}
\usage{
plot_qqline(
  data,
  ycol,
  group,
  symsize = 3,
  symthick = 1,
  s_alpha = 1,
  ColPal = "all_grafify",
  ColSeq = TRUE,
  ColRev = FALSE,
  TextXAngle = 0,
  fontsize = 20,
  Group,
  ...
)
}
\arguments{
\item{data}{a data table e.g. data.frame or tibble.}

\item{ycol}{name of the column containing the quantitative variable whose distribution is to be plotted}

\item{group}{name of the column containing a categorical grouping variable}

\item{symsize}{size of symbols, default set to 3}

\item{symthick}{thickness of symbol border, default set to 1}

\item{s_alpha}{fractional opacity of symbols, default set to 1 (i.e. maximum opacity & zero transparency)}

\item{ColPal}{grafify colour palette to apply, default "all_grafify"; alternatives: "okabe_ito", "bright", "pale", "vibrant", "contrast", "muted" "dark", "light".}

\item{ColSeq}{logical TRUE or FALSE. Default TRUE for sequential colours from chosen palette. Set to FALSE for distant colours, which will be applied using  \code{scale_fill_grafify2}.}

\item{ColRev}{whether to reverse order of colour choice, default F (FALSE); can be set to T (TRUE)}

\item{TextXAngle}{orientation of text on X-axis; default 0 degrees. Change to 45 or 90 to remove overlapping text}

\item{fontsize}{parameter of \code{base_size} of fonts in \code{theme_classic}, default set to size 20.}

\item{Group}{deprecated old argument for \code{group}; retained for backward compatibility.}

\item{...}{any additional arguments to pass to \code{ggplot2}\link{stat_qq} or \code{ggplot2}\link{stat_qq_line}.}
}
\value{
This function returns a \code{ggplot2} object.
}
\description{
This function takes a data table, a quantitative variable (\code{ycol}), and a categorical grouping variable (\code{group}), if available, and plots a QQ graph using \code{stat_qq} and \code{stat_qq_line}.
}
\details{
Note that the function requires the quantitative Y variable first, and can be passed on a grouping variable as \code{group} if required. The graph plots sample quantiles on Y axis & theoretical quantiles on X axis. The X variable is mapped to the \code{fill} aesthetic in\code{stat_qq} and \code{colour} aesthetic for the \code{stat_qq_line}.

Colours can be changed using \code{ColPal}, \code{ColRev} or \code{ColSeq} arguments. Colours available can be seen quickly with \code{\link{plot_grafify_palette}}.
\code{ColPal} can be one of the following: "okabe_ito", "dark", "light", "bright", "pale", "vibrant,  "muted" or "contrast".
\code{ColRev} (logical TRUE/FALSE) decides whether colours are chosen from first-to-last or last-to-first from within the chosen palette.
\code{ColSeq} decides whether colours are picked by respecting the order in the palette or the most distant ones using \code{\link[grDevices]{colorRampPalette}}.
}
\examples{
#Basic usage
plot_qqline(data = data_cholesterol, 
ycol = Cholesterol, group = Treatment)

#with faceting
plot_qqline(data = data_cholesterol, 
ycol = Cholesterol, group = Treatment, 
fontsize = 10)+facet_wrap("Treatment")

}
