% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qqline.R
\name{plot_qqline}
\alias{plot_qqline}
\title{Plot quantile-quantile (QQ) graphs from data.}
\usage{
plot_qqline(
  data,
  ycol,
  xcol,
  symsize = 3,
  symthick = 1,
  fontsize = 20,
  alpha = 1,
  ColPal = "all_grafify",
  ColRev = F
)
}
\arguments{
\item{data}{a data table e.g. data.frame or tibble.}

\item{ycol}{name of the column containing the quantitative variable whose distribution is to be plotted}

\item{xcol}{name of the column containing a categorical variable}

\item{symsize}{size of symbols, default set to 3}

\item{symthick}{thickness of symbol border, default set to 1}

\item{fontsize}{parameter of \code{base_size} of fonts in \code{theme_classic}, default set to size 20.}

\item{alpha}{fractional opacity of symbols, default set to 1 (i.e. maximum opacity & zero transparency)}

\item{ColPal}{grafify colour palette to apply, default "all_grafify"; alternatives: "okabe_ito", "bright", "pale", "vibrant", "contrast", "muted" "dark", "light".}

\item{ColRev}{whether to reverse order of colour choice, default F (FALSE); can be set to T (TRUE)}
}
\value{
This function returns a \code{ggplot2} object.
}
\description{
This function takes a data table, X and Y variables, and plots a QQ graph using \code{\link[ggplot2]{ggplot}}.
}
\details{
The function uses \code{stat_qq} and \code{stat_qq_line} geometries (get help with \code{?stat_qq_line}).
Note that the function requires the quantitative Y variable first, and groups them based on an X variable.
The X variable is mapped to the \code{fill} aesthetic in\code{stat_qq} and  \code{colour} aesthetic for the \code{stat_qq_line}.
ColPal & ColRev options are applied to both \code{fill} and \code{colour} scales.
}
\examples{
#Basic usage
plot_qqline(data_cholesterol, Cholesterol, Treatment)

}
