% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qqmodel.R
\name{plot_qqmodel}
\alias{plot_qqmodel}
\title{Plot quantile-quantile (QQ) graphs from standardised residuals of linear models.}
\usage{
plot_qqmodel(model, ...)
}
\arguments{
\item{model}{name of a saved model generated by \code{simple_model} or \code{mixed_model}}

\item{data}{a data table e.g. data.frame or tibble.}
}
\value{
This function returns a \code{ggplot2} object.
}
\description{
This function takes a linear model (simple or mixed effects) and plots a QQ graph using \code{\link[ggplot2]{ggplot}} after running \code{fortify} or \code{fortify.merMod} to calculate standardised residuals.
}
\details{
The function uses \code{plot_qqline}, which relies on \code{stat_qq} and \code{stat_qq_line} geometries (get help with \code{?stat_qq_line}).
The graph plots standardised residuals from the model on Y axis & theoretical quantiles on X axis.
Other
}
\examples{
#Basic usage
m1 <- simple_model(data_2w_Festing, "GST", c("Treatment", "Strain"))
plot_qqmodel(m1)

}
